2019 Refactor Intcode VM
Replace input channel with an input function again, while also providing a helper to make a function from a channel. This is slightly clunkier for pure channel users but allows use cases where the client needs to know exactly when an input is needed (such as Day 13) and compute the answer as late as possible.