2019 Rewrote Day 14 Part 2
This is more like what I was trying to do initially. This is guaranteed to be correct as it deals only with ints. That being said, "need" is a bit weird since it says how many need to be created, not how many are left/used by the reaction. Maybe I can rework this even more?