2019 Day 18 Part 1
I guess this is a constrained traveling salesman problem? This was brutal!

I went down so many rabbit holes, and cleaning this up took quite a while (since I wanted to cut out as much hackiness as possible) but this is a really good, general solution now. Converting mazes to graphs is definitely a nice savings, while still allowing for changing availability cleanly as doors get unlocked.