2019 Refactor Day 19 Part 2
There's no need to check the entire 100x100 block to see if it is good, we can *just* check the right side. The tractor beam always has full, connected lines, and we know that the starting x is monotonically increasing, so ensuring that the right side stays at blockX+99 for all 100 rows is enough to prove that the 100x100 square is complete.

Also, start the search at y=100. No valid block can possibly be present at y=50!