2019 Day 20 Part 2
That wrench of inner and outer was annoying to parse right...at least I got it. But man, I had *so* many stupid bugs to fix in this part...

Also, I probably should write a generic BFS implementation (maybe even A*) instead of having to dig into the guts of one every time.

Also also, there *has* to be some graph analysis thing to prove that some inputs are unsolvable for part 2. This approach just loops forever in that case!