2020 Day 4 Part 2
Again, indecision cost me. But even more was me using chained comparison operators, but incorrectly! a > b < c (which I first tried) is the same as b < min(a, c), which is certainly *not* what I wanted. What I really needed was to *negate* the full valid condition (a <= b <= c) by just wrapping it in parens and putting a "not" at the front! *sigh*