2020 Day 7 optimization
Okay, it's the morning, and I'm feeling (slightly) less annoyed. I could have leaderboarded with the naive solution for part 2, but in my defense my code works for intense inputs. So I decided to prove that and construct an intense input that kills the naive solutions, but my part 2 (and a similar one for part 1) handles fine.

(Admittedly memoization is probably simpler to think through, but still...)