2020 Day 10 Part 1
I absolutely bombed this. My knee-jerk reaction of just sorting the list was spot on, but I didn't notice that I forgot to update my last seen number. This meant that my assert that the difference was in (1, 2, 3) failed! I got sidetracked because of this and was thinking about hamiltonian paths and graph theory and crap. What a waste of time.
Furthermore, I missed adding the charging outlet joltage (0) to the list of numbers, wasting another minute after I finally fixed the above idiocy.
An absolutely dreadful showing today.