2020 Day 13 Part 2
Well this made me exercise my brain for sure. If it was asking for the first time where all the busses arrived at the same time it'd be an easy least common multiple problem, but the offset times threw me. For lack of any better approach I just worked through how to optimize the naive search until it was sufficiently fast. My theory was that "merging" two busses and treating them as one (by knowing when they first line up properly and the cadence between when they line up again) would sufficiently optimize the search and let me complete the search in a reasonable amount of time. Thankfully it panned out, but man is my code scuffed!