2020 Refactor Day 13
Between looking at the subreddit and pausing to think through my code, I can clean it up a good bit. It helps to observe that the cadence increases by multiplying the current cadence by the bus cadence.
I can also express the next available time for a bus in a closed form expression without looping. (I knew I could while solving, but I didn't want to spend the time to think through it!)

Also, a bunch of comments in the subreddit mention a Chinese remainder theorem. I should look into that and see what's there (and whether I accidentally stumbled across some of the relevant logic for it).