2020 Day 14 Part 2
It took me longer than I'd like to have thought through how to deal with the floating bits, but I managed to write something working on the first try. (In fact, when my initial test of it broke it was in fact me passing the wrong parameters!) I could have done this part a bit faster if I tested it properly the first time (and if I didn't accidentally discard the floating number in the part2 code by assigning it to a "quantum" variable that wasn't used), but I don't think I'd have saved much time overall. (Certainly not enough to leaderboard.)