2020 Day 16 Part 2
This was an interesting extension to the problem. I was scared that I'd have to determine the order of the ticket fields, and lo and behold we do have to do that. But as it turns out, determining the candidates for each field and doing a naive backtracking search for an order that works runs in a small enough amount of time to be feasible. I'm sure that determine_rule_order can be optimized *way* more, but it would have absolutely cost me the leaderboard to do that! (That being said, I was worried for a moment when it took a few seconds to terminate when I first ran it!)