2020 Refactor Day 16 (part 2)
Eliminate the recursive backtracking method in determine_rule_order in favor of the greedy approach to assigning rule order. At first it may not seem obvious, but the greedy approach is guaranteed to find a solution if it exists. If this fails to find a solution then that means that two (or more) fields could all fit the same two (or more) possible rules with nothing to disambiguate them. Even if we were to use other methods (like choosing rules that can now only fit one possible field position) this fundamental ambiguity will prevent us from uniquely assigning the rule order!