2020 Day 19 Part 2
This was definitely an interesting extension. My code is definitely not prepared for arbitrary grammars, but I did at least eventually figure out how to generate bounded possibilities (so many dumb bugs there) and how to simplify the messages/rules so that recursion ran in a reasonable amount of time. (But man, this hurt!)