2020 Rewrite Day 19 Part 1
In the subreddit I saw lots of regex, and one of my earlier thoughts was regex. But at the time I couldn't think of how to figure it out! It turns out that it really isn't that bad, though.

Too bad I didn't think of recursively generating the regex. This was much simpler to write *and* it runs super fast!