2020 Rewrite Day 19 Part 2
It took some doing, but I did come up with my own solution for Part 2 using regex. I was of course somewhat influenced by looking at the megathread, but the main important observation is that rule 8 is <rule_42>+ and rule 11 is <rule_42>{n}<rule_31>{n}. If you're willing to abuse the regex engine you can get a working solution this way!