2020 Refactor Day 19 Part 2
Now that I wrote it "my way", I can implement (and document) the super smart way to do it. Due to the properties of rules 0, 8, and 11, rule 0's regex ends up being (<rule_42>+)(<rule_31>+), so long as the first group is longer than the second group! This is by far the nicest solution in my opinion. (In fact, nearly half of my part2 function is just an assertion verifying that the assumption holds, and comments explaining the cleverness!)