2020 Refactor Day 21
It turns out that the mapping of allergen to ingredient computed in Part 2 didn't even use the possible_bad_ingredients set from Part 1. As such, it's pretty much cleanest to determine the allergen_to_ingredient mapping for both parts and then use that mapping to answer the Part 1 and Part 2 questions. (In fact, I don't *really* need the list of safe ingredients for Part 1!)