2020 Day 22 Part 2
I'm actually really pissed off about this. A critical piece of the problem was *not* highlighted (namely that only n cards of the deck get copied for the sub-game) causing me to waste an insane amount of time. I should have been able to easily leaderboard, probably even top 30. But instead because of that nonsense I ranked 1168 which is absolutely abysmal for me this problem.

Whatever.