Refactor mod_mult_inv
It turns out that in Python 3 pow(base, -1, exp) returns the modular multiplicative inverse. Goodbye extended Euclidian algorithm, hello built-in Python function. (Both because it's faster, and because it means I won't get confused looking at the code. I need to be more familiar with the extended Euclidian algorithm, but this lets me punt the issue!)