Change naming convention of helper libraries
"helpers." is a lot to type. Using "lib." as the prefix is much less to have to type when I'm in a rush solving a problem, so it's a better choice.

Additionally, rename the advent of code io library to "aoc" instead of "input" since it can also submit answers.