2021 Day 3 Part 2
I wasted more time on Part 2 than even on Part 1! Pulling out determine_common was obvious enough, but I skimmed over the rules too quickly to see the requirement to have a fallback when the counts are even. Moreover, I was passin an int default instead of a string, and other dumb mistakes! Oh well, at least it worked in the end.