2021 Day 4 Part 1
Protip: Don't be dumb when solving problems. If the problem asks for the score of the *winning* board, make sure that you actually give *that* score, and not the score of some random board...

I mean, the problem itself wasn't that bad. I actually wrote a class for this, amazingly. But *wow* was that bug dumb, and wow did it cost me a ton of time. I SOLVED PART 1 MANUALLY IN NOTEPAD BOTH TO GET AN ANSWER AND TO HELP DEBUG!