2021 Day 6 Part 2
Yep, this was an optimization problem all right. I had a feeling that I was going to be punished for doing Part 1 the dumb way!

I did get to exercise collections.Counter here though. Unfortunately the total method is only available in 3.10+ and I'm running 3.9, so I wasted some time on that on top of having to rewrite Part 1 the right way to verify that I was doing it right!