2021 Day 7 Part 1
There's got to be smarter ways to search for the optimal peak, and I know this code is bad, but speed is key. And even still with an answer in in 4 minutes I'm rank 864! That's insane! (I really typed slow honestly...)

The worst part was me wasting time thinking of how to go at it more optimally instead of brute forcing the options! (And I spent time getting stats about the range to see if brute force was viable which cost time!)