2021 Day 8 Part 2
This makes my bungling of Part 1 look tame! Really this is just a pruned backtracking search, but it took me way too long to grok through how to do it! (Not to mention the right form of my valid pattern mappings!)

It works well enough in the end, but holy crap did I take too long...