2021 Rewrite Day 7 using pseudo-binary-search for slopes
This isn't the *most* efficient approach for Day 7 (part 1 is the fuel cost of the median of the positions and part 2 is a fuel cost near the mean of the positions), but it's dramatically more efficient for large inputs (such as the 1 million position input from the big inputs thread). Having a find_continuous_curve_minimum function available may be helpful in other problems in the future.