2021 Rewrite Day 10
I really should have used some lists and generic handling for my validation code. Fewer characters and less to write, as this rewrite demonstrates!

(I also used some list comprehension to shrink the part1 and part2 code, but that's more complex to think through on the fly so no biggie that I didn't do that the first pass.)