2021 Day 12 Part 2
This is an interesting variation. Too bad I messed *this* up royally as well. I read the problem as allowing double-visits to *all* small caves...which was taking forever to run. So I ran it on the example and got a totally different path count! This led me down a giant rabbit hole until I realized that only *one* double visit is allowed, total!

I really need to stop misreading the problem, it's costing me an incredible amount of time. (Plus I need to make fewer idiotic mistakes...)