Write OCR lib for Advent of Code text
This is not the first time that we've had to parse text from a grid of dots/markings in Advent of Code. It's high time I have a programmatic solution for this!

Since Advent of code uses the same fonts year to year (at least for the same character dimensions) this is a much simpler problem than the general OCR problem. I simply need to isolate characters in the stream (looking for blank lines delimiting breaks between characters), extract the dimensions and bit pattern representing that character, and look that character up in a cache. (And if I don't know the character yet, ask the human in the loop!)

Credit goes to multiple people in the Reddit for inspiring me to do this, finally. (And also for having lists of known characters so that I can seed my OCR lib with extensive knowledge without having to go back to *every* previous Advent of Code problem which required converting images to text!)