2021 Day 14 Part 2
So this *was* an optimization problem! You know what? I'm not even mad about not leaderboarding and dropping to 287. It took me a bit to realize that I could just memoize the results of the rules after N iterations but that can get the answer *quick*. The only tricky bit is not double-counting elements in the expansion, which I *did* get wrong a couple times.

I'll clean this up in a bit, there's no reason why this couldn't be done for Part 1 as well.