2021 Day 15 Part 2
Oh, now that's an interesting twist. It took me a little bit to think through how to neatly stitch together a bigger grid, but then I realized that my grid library can convert *from* a dictionary too so I abused that!

Then it was just waiting for Dijkstra's shortest path to finish running. I really should add an optional heuristic to dijkstra_length to turn it into A* to speed things up...