2021 Drastically Optimize Day 17
Since I couldn't do the problem well when it came out, I guess I got super into optimizing it. It now runs fast even on ridiculously big inputs.

Part 1 is simple, the maximum possible y velocity is known and computing the max height of the parabola is simple.

Part 2 is much more interesting. It's possible to find the x and y solutions independently, as well as the ranges of how many steps each component takes to reach the target. This has been *heavily* optimized to reduce redundant work by jumping ahead where possible since the low and high ends of the step ranges are monotonically increasing for each solution.
Once this is done it's possible to loop through the y solutions in reverse and keep track of how many x solutions enter and exit the step range for the y solution. Thus we can quickly count all unique solutions of starting x and y velocities.

Also I used parse.parse to parse the input. I was wanting to do that when the problem came out but I bungled that too :/