2021 Day 18 Part 1
LET'S GOOOOOOOOO! RANK 32!!!!!!

I thought that I was *so* slow doing this, but apparently I did well! try_explode is a hot mess, but it works at doing the explode step. I kept testing that against all inputs, but holy crap it is a nasty mess of string logic. I'm sure there's a better way to do this, but I did what I could think through.

At least try_split was pretty simple recursion, reduce was a simple loop, and magnitude was simple recursion as well. Really the explode process was where all the complication was!

(Oh, and yes I am using eval to convert to a Python list and str to convert it back to a string for the super nasty try_explode logic. Yes this is horrible practice. I don't care.)