2021 Day 18 Part 2
RANK 30 WOO!!!

I'm not sure if there's really any better way to do this other than testing all the pairs...so that's what I did. I was worried that my reduce logic would be too slow (and it isn't exactly fast), but it did give me an answer in ~5.5 seconds which is plenty fast for a first pass!