2021 Refactor Day 18
The string-based explode implementation was a horrid mess. I took inspiration from someone on the Reddit and rewrote it to not be a hot mess, and not only is it easier to understand but it is also a good bit faster!

While I was at it I cleaned up some naming, made all the code consistently use isinstance(num, int) at the beginning for type checking, simplified the input parsing step, inlined the add function (because it's trivial), and changed part 2 to use a generator expression. All in all much cleaner, though it's about the same length.