2021 Further refactor Day 18
Since all explosions are done before any split is is possible to make a single pass to execute all explosions, with careful bookkeeping. We just need to keep track of any right-propagating values and pass them along, really. This actually shortens the explode function as well as speeding it up!

I also tried to optimize split_num but while I did make it a little faster, it also became an unruly mess. There may be a clever way to make it faster without making the code sinful, but I'm satisfied for now.

Also, use json.loads on the input for more safety.