2021 Refactor Day 19
I very much demonstrated a need for a graphics concept library in my helper libs. lib.graphics aims to fill that void with Point3D/Vec3D and Mat3D. Not all concepts are present (rotation around arbitrary axes isn't present and no native 2D support is there yet for example) but this is a good start. It makes the Day 19 solution code much cleaner.

Just in case, this has been added to the day template for future problems.

I also went ahead and reworked the solution code a bit. I actually did that before working on lib.graphics and it turns out that that's good, lib.graphics was 5-6x slower with the original algorithm despite my efforts to speed it up. (There may be more I can do to improve it, but I've done enough for now!)
1) Instead of regenerating the alternate orientations constantly, generate them once and keep them around
2) Match scanners based on individual known scanners. This keeps the list of coordinates to cross-examine down significantly.
3) Use collections.Counter to count matched offsets. If a pair of scanners as 12 or more hits on the same offset then that offset is a winner! This does the same job as the old algorithm but with much less complication. Plus it ends up being faster!
4) Deduplicate the solve code from part1 and part2. That was just wasteful duplication.

It's shorter, faster, *and* clearer now!