2021 Refactor Day 20
I wanted to use a set for the live pixels in the image, but thanks to the alternating pattern and nontrivial algorithm it gets kind of nasty. Plus it slows down the code!

However, I can pre-translate the whole string to 0's and 1's so that the index determination step is easier. Plus the iteration can be pulled into its own function making part1 and part2 super basic. Add some inlining and list comprehension/generators and the code gets pretty nice!