2021 Optimize Day 19
It's possible to prune out a lot of inappropriate scanner pairs by comparing their beacon distance counts. If two scanners align then they must have 12 or more beacons which have the same Manhattan distances between them. This is quick and easy to check with collections.Counter.

A few more smaller improvements were made by inlining some code and reorganizing it, but on the whole the algorithm is the same as before. The code is really quite nice now though!