2021 Refactor Day 21
Inline the game logic into part1 since it's not really reused in part2. It also can use a die variable instead of a generator which feels marginally nicer.

Also, rework part2 a good deal. I didn't like the triple nested loop, or the if switch for the active player. Instead I can precompute the moves (and universes in which those moves happen) as well as use tuples/lists to track player positions, scores, and win counts. This makes the count_wins function much shorter, and also means that it can support 3 player games with no code change! (The code runs much slower for 3 players though...)