2021 Day 22 Part 2
WHAT?! I just got rank 18!!!!! Are you serious?! WOW!

Now this was an interesting twist, and I was wary of it once I saw the problem I hit in part 1. I went through a few ideas quickly which sounded hard to figure out, but then I realized that I could simply count the tiles turned on for each command so long as no subsequent command touched that region. Then even more importantly I realized that count_uninterrupted could help me determine the overcounting that it was doing itself! I have no idea if this is the "right" way to do this (I've never thought about a problem like this before) but it clearly worked well!