2021 Refactor Day 22
Now that I have some time to breathe, I can make part 1 follow the same approach. It isn't really any more efficient, but it is nice to have one solution path. The only thing is that the ranges need to be clipped to the window of interest for part 1.

Also, clean up some naming and clean up function logic.