2021 Day 23 Part 1
This was abysmal. Only now that I finished part 2 do I realize how simple this problem actually was. I didn't read the rules carefully enough to notice that the creatures don't move from the hall until they have a place to go. That *dramatically* simplifies the search space.

Frankly I got lucky to even get the right answer. And even still this was an abysmal rank 765...