2021 Day 23 Part 2
Yeah, so I tried so long to optimize the search space for the wrong problem. I tried better heuristics, better neighbor functions, all sorts of crap. None of it was right because I was trying to solve the wrong problem...

As soon as I implemented the correct neighbor function it solved quickly. Even without a heuristic. Oi...

This landed me with an even *worse* rank 980, after an insane 2 hours 42 minutes!