2021 Day 24 Part 1
Holy crap. That was brutal. I went down the wrong tracks for literally 6+ hours.

I still don't know how to solve this programmatically without analyzing the input. I tried simplifying expression trees like a compiler optimizer (which is hard), doing attempts at "search all possibilities" which failed hard, etc. Nothing worked.

In the end after seeing the repetition of groups in the input I spent I'm not sure how long analyzing the function block, decompiling it into Python, then realizing that this could be solved with a stack.

This is an interesting problem, but I think it's too much for Advent of Code. Either that or I'm missing some super good techniques...