2021 Rewrite Day 24
This was a HUGE undertaking, but incredibly rewarding. I now have a programatic solution which doesn't make any major assumptions about the input structure! It actually runs the ALU and finds which digit substitutions are valid by combining symbolic expressions and branching on the eql instrutions!

I'm sure this can be faster with a symbolic expression library not written by me, and there are probably other things that can be improved. But this actually works! YAY!