2021 Refactor and Optimize Day 24
The generalized symbolic expression approach is much nicer conceptually, but the way it was implemented was not exactly fast. The biggest culprit was the excessive use of substitution options for expressions and terms, both for checking equalities and for generating the final answer. This is actually fairly easy to improve though. Keeping track of the input domains of symbols and output domains of expressions and terms allows for suitably accurate simplification of expressions as well as checking whether equalities and inequalities are plausibly satisfiable. This along with generating solution constraints instead of substitution options in run_alu makes run_alu run much, much faster.

This does require some extra logic to find the solution at the end, but a simple backtracking search over the inputs (pruning the search when any constraint can no longer be satisfied) works well enough. This could probably be optimized by only substituting symbols in constraints which contain those symbols, but the find_solution step doesn't take long at all. Such optimization would likely only make sense if trying to generate all solutions.

With this the programmatic solver now takes ~35-40ms to find each answer, from scratch, truly with no assumptions about the input! (Only some unhandled cases for expression composition.)