Add a performance library
Sometimes I want to time how long a solution takes. I've been doing this manually but that's a pain to do. A barebones performance library with a timed decorator will make this much easier in the future.

Down the line I may add other facilities to this to help me profile and optimize solutions.