2015 Rewrite Day 11
Converting fully to an integer to increment then back to a string then back to a list of numbers feels so wasteful. It's much nicer to convert to a list of numbers, increment that form, check that form, and convert back to a password once all the checks pass.

Plus with a single next_legal_password function the part1 and part2 functions become much simpler!