2015 Day 19 Part 2
I already knew the "trick" to this one, so I wanted to try to solve it without applying *all* of the trick. I didn't succeed though, and had to fall back on just molecule counting.

This really isn't a satisfying solution, at some point I really should finish that cyk parsing algorithm and use it to solve this generically rather than this nasty special-case garbage...

Time: 1:04:27.734970