Implement a CNF/CYK library for grammar problems
The 2015.19 solution is not particularly satisfying, and the 2020.19 solution is pretty clever but makes some assumptions about the lengths of generated strings. A full CYK parser is a heavy hammer to throw at these problems, but it is very general-purpose and works without assumptions.

Admittedly the new solution code is slower than the original solutions, but being general-purpose it may be useful in the future. More importantly, it just feels like a better way to solve these problems!