2016 Day 2 Part 1
Oof, that's a harder Day 2 too! Still not too bad, but moving around in a fixed grid while bounding the coordinates was interesting.

Honestly though, I wasted a ton of time in mapping the end coordinates to the keypad. Doing it manually is dumb, but my brain wasn't immediately processing how to do it smart. (Which is silly, because isn't it just 3*y + x + 1?)

Time: 0:03:12.710717