2016 Day 11 Part 1
Okay, that was kind of ridiculous. Believe it or not this would have been 59th on the leaderboard for this year!

The idea here is simple: Do a breadth-first search of the state space. The problem is that generating neighbor states is kind of ridiculous, 86 lines of code! Not to mention the complicated input parsing.

This really doesn't seem like a good middle-of-the-month problem...

Time: 1:02:39.874640