2016 Day 11 Part 2
At least Part 2 wasn't bad. I had to spend time pulling code out to be shareable and adjusting the initial state, but then it was just a waiting game.

Seriously. This is not fast. Either I'm missing some very good insights or this problem was...bad...

Though this would have been 42nd on the leaderboard, somehow.

Time: 1:10:13.799028