2016 Rewrote Day 18
The Reddit had some clever observations that I missed in the heat of the solve. The center tile doesn't matter, and in fact a tile is a trap iff the left and right tiles don't match. Since that's an XOR operation, some careful bit operations can solve this problem much faster. (Though I do have to bring in gmpy2 to speed up the counting of trap tiles.)