2016 Day 19 Part 1
Okay, so a linked list problem. I of course tried doing this with a list at the beginning but that gets bogged down with all the shifting. I wonder if I should have a linked list class in my libraries somewhere to not need to write one from scratch in the future? Meh...

Time: 0:06:54.566068