2016 Day 19 Part 2
Oh now *that* is an interesting twist! Took me a little bit to think through how to approach it, but we don't actually care who is stealing, we care who is removed! Some careful tracking of what the current target is makes this not too bad to solve.

Well...if you actually move *left*. I wasted about 1 minute 45 seconds because I was moving to the right! That bumped me from what would have been a second place time to a third place time :(

Time: 0:15:57.492390