2016 Day 20 Part 1
Well this was interesting to think through. If I deal with sorted ranges then I can easily keep track of the next candidate IP. The first candidate IP which *isn't* blocked by a range must be the lowest, otherwise the next candidate is either the current candidate *or* one past the current test range's high value, whichever is greater.

This would have been first on the leaderboard!

Time: 0:02:08.236847