2016 Rewrite Day 21
Since most of the inverses are either the application operation or a very simple twist on the application operation a lot of part1 and part2 can be deduplicated. Only the "rotate based on" operation has a complicated inverse.

That being said, even that can be improved with some preprocessing. For passwords of length 8 only one possible inverse can exist so everything is invertible, but some special handling can pretty easily detect if there are conflicts that can't be handled.