2016 Day 23 Part 2
Oh now there's a twist! Thankfully this was explicit about needing to do some decompiling of the instructions (as well as what to look for), but pattern-matching the multiplication loop was still a bit tedious. I *was* going to complete the validation but punted that to get my answer in.

This would have been 13th, not bad.

Time: 0:27:53.074428