2017 Day 3 Part 1
There's probably a more efficient way than actually generating the coordinates in order, but this seemed like a good way to prepare for part 2 and wasn't bad to think through.

I *should* have been a minute faster, but I was using variables in my loops in gen_coords which messed up the coordinates and of course my answer :(

Would have been 27th.

Time: 0:04:55.187432