2017 Day 13 Part 2
I initially was thinking that this would be a Chinese Remainder Theorem problem, but we're given *incongruencies*, not congruencies! There's probably a way to do this more efficiently, but I ended up just brute forcing it.

Again though I messed up the mod arithmetic and was doing the wrong offsets! Bah!

Would have been 42nd.

Time: 0:15:32.504628