2017 Refactor Day 13
Parts 1 and 2 share some logic in determining the cycle length which can be pulled into parse_input rather than duplicated.

Furthermore, Part 2 *can* be done using a variant of the Chinese Remainder Theorem. It's a bit more complicated since we're dealing with incongruencies rather than congruencies, but with some careful logic and management the incongruencies can be converted to congruence options and those options can be checked all at once. There is a risk of exponential blowup, but the new lib.math.chinese_remainder_incongruence function takes care to minimize the risk of exponential blowup and, if it does happen, minimize the impact.

This was a fruitful exercise as well since some improper logic in lib.math.chinese_remainder was also uncovered when the accrued n and new mod *both* need to be divided to make the congruencies coprime.