2017 Day 14 Part 1
I was not expecting knot hash to come back! I was confused for a bit since the previous problem (day 10) had talked about part 1 being a  "single round" but that doesn't generate 128 bits! I had to make an educated guess that part 2 was the real knot hash, which it was. That still required an annoying amount of updating to my logic to make it work though.

At least once that was working I just needed to count on bits!

Would have been 43rd.

Time: 0:06:58.953101