2017 Day 14 Part 2
Conceptually this is a simple flood fill algorithm built on top of the grid generated in part 1. Why did it take so long then? Because I forgot that the output of bin would be missing some leading zeros leading to the rows being shifted and my on_positions set being incorrect!

Would have been 22nd.

Time: 0:15:01.774260