2017 Day 15 Part 1
I wasted a *ton* of time on this. I was thinking that at 40 million pairs I'd have to analyze the sequence or something, but it turns out that brute force is the way to go.

Then of course once I did this the right way (brute force) I accidentally generated the numbers incorrectly by passing the a seed into the b generator...what a waste...

Time: 0:15:27.519833