2017 Optimize Day 17 Part 2
Since we only care about numbers when they're inserted after 0 (which is fixed at index 0) we can skip ahead whenever insertions don't loop around. This drastically improves the runtime of the problem.

While implementing this I also realized that my original solution had a bit of an oversight if target_pos was not 0. target_pos would shift if the insertion position was less than or equal to target_pos, not only if it was equal to target_pos! Thankfully that doesn't matter for the target being 0.