2017 Day 18 Part 2
Holy crap everything fell apart here! I thought this would be a fun little coroutine problem, but nooooo.

I started out writing a simple Python coroutine that I'd instantiate twice and run, but it was running forever. Dismayed I then thought that maybe this was an icky reverse engineering problem. It isn't, but I still spent a ton of time reverse-engineering the code to find that it's a bubble sort of 127 values.

So I tried counting the steps for bubble sort for the generated sequence. It sort of worked, at least in confirming on the site that that was the right ballpark. However, the numbers I got were wrong. (Don't really know or care why.)

Then I reimplemented the *entire* thing with a class for easier debugging. Surprisingly the number of values passed between them did *not* stay at 127, so something was clearly going very wrong!

After *way* too long debugging I finally found two really, really stupid bugs:
1) When parsing the instructions I forgot to convert parameter 0 of jgz to int if it wasn't a letter, breaking the program.
2) I also implemented jgz with a simple equality check against zero, not checking for greater than zero!

Honestly, if I'd have done those right initially I'd have saved so much time. This was an absolute disaster.

An hour past the end of the part 2 leaderboard :(

Time: 1:40:10.073833