2017 Rewrite Day 18
I took the opportunity to write a go-style channels library and rewrite the solution to this using threads and channels. While kind of overkill for this problem, it is a good test case for channels which I know I will want when re-solving 2019's IntCode problems.

As an added bonus, part 2 involves detecting deadlocks which I know will also come up in at least one 2019 problem so I have this handled already!

The documentation isn't really *great* in the channels library, but it's good enough for me.