2017 Day 20 Part 1
I have a feeling that I did not do this as intended. I did analysis and figured that the particle accelerating the least with the lowest starting velocity would have the least velocity over time and would tend to be nearest the origin in the long run.

Really I needed to test a few more assumptions to be sure with my approach, but I think this may have been more complicated than was intended...

Wouldn't have been top 100 :(

Time: 0:13:49.579878