2017 Rewrite Day 20
I didn't feel right that my solution was so full of assumptions. It's actually possible to solve both part 1 and 2 deterministically without assumptions using quadratics.

Starting with part 2, candidate collision times can be found by analyzing just the x coordinate quadratic formulae of all particle pairs. These are not guaranteed to produce collisions, but they are a prerequisite. Once all candidate integer times which produce x coordinate conflicts are found it is simple to iterate through them (in ascending order), find where the particles will be at that time, and eliminate any collisions.

Part 1 ends up being the more interesting one in this approach. It is possible to construct a long-term manhattan distance quadratic based on the component formulae, but care must be taken to negate any component formulae which will tend towards the negative axis. However, once this is done we simply need to find the quadratic which grows the slowest.