2017 Day 21 Part 1
Holy cow did I botch this. I mistakenly thought that it was going to alternate between the 2x2 rules and 3x3 rules and figured it was going to be a memoization problem in part 2. But no, it does a 3x3 rule, then 2x2 rules, then 2x2 rules, then back to 3x3 rules.

This is because 3x3 goes to 4x4, which goes to 6x6 (still divisible by 2), then to 9x9. We have to divide 2 out *twice*!

After a *lot* of flailing about trying to figure out what my bug was I finally realized that my enforced alternation was the problem. It took some time to rewrite it to break apart a grid into chunks and reconstitute a new grid, but that worked first time.

(The annoying part is that that was my first instinct, but I thought I was being clever by introducing memoization!)

Would have been only 10 minutes after the leaderboard closed, somehow.

Time: 0:51:05.174428