2017 Day 23 Part 2
Ew. This is an honest to goodness reverse engineering problem, unlike the previous ones which could easily be solved with some VM optimization of certain patterns.

This ended up taking me two passes to reverse engineer because I made some dumb errors in the first pass. I actually didn't 100% reverse engineer it, but I got it far enough that it gave me the answer which is the point of the exercise.

Would have been 68th on the leaderboard, at least.

Time: 0:44:42.572933