2017 Day 24 Part 1
Well that wasn't bad! This is essentially a pathfinding problem. (In fact, it *is* a pathfinding problem now that I think of it, I just did it manually!)

My main worry was exponential complexity due to repeated iteration over the part list, but a quick port to candidate indices lookup fixed that really fast :)

Would have been 7th!

Time: 0:06:42.833727