2018 Refactor Day 3
Turning the ranges into coordinate lists requires the same for loops 3 times, most notably with the icky range. Taking care of that inside the parse function is preferable.

That being said, the end code length is nearly the same. Furthermore, adding collections.Counter objects seems to slow things down, so the code is still manually counting hits. Nonetheless, this does seem a bit cleaner to me.