2018 Day 4 Part 1
Oof. That was a lot to unpack for day 4! I got tripped up a few times in the parsing, but even moreso in managing the data.

Initially I thought that the sleep times could extend long periods of time and that I'd need to use the datetime library, but all sleeping being during the midnight hour makes things so much easier. Still a lot to track though.

Would have missed the leaderboard by 89 seconds. But with how much parsing there was, I'm fine with that.

Time: 0:19:10.778197