2018 Day 5 Part 2
So now we have 26 candidate strings and need to fully reduce all of them. In other words, my inefficient part 1 approach is not going to cut it.

Thankfully this can easily be done with a queue and stack since any reaction just "exposes" the previous unit in the polymer. It took a bit longer than I'd like to have coded up the smart version though.

Would have been 96th or 97th. Ouch! I'm surprised I dropped down so low, but that's what I get for doing part 1 so suboptimally.

Time: 0:10:11.151292